# config/swagger.yml
swagger: "2.0"
info:
  title: API Python utilizando Flask
  description: Projeto Python utilizando Flask e SQLServer.
  version: 1.0.0
host: api.example.com
basePath: /v1
schemes:
  - https
paths:
  /clientes:
    get:
      properties:
        id:
          type: integer
        name:
          type: string
        # Both properties are required
        required:
          - id
          - name
      tags:
        - Clientes
      produces:
        - application/json
      responses:
        200:
          description: OK
  /cliente:
    post:
      tags:
        - Clientes
      produces:
        - application/json
      responses:
        200:
          description: OK
    put:
      tags:
        - Clientes
      produces:
        - application/json
      responses:
        200:
          description: OK
  /cliente/{id}:
    get:
      tags:
        - Clientes
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
          minimum: 1
          description: Parameter description in Markdown.
      responses:
        200:
          description: OK
    delete:
      tags:
        - Clientes
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
          minimum: 1
          description: Parameter description in Markdown.
      responses:
        200:
          description: OK

  /pagamentos:
    get:
      tags:
        - Pagamentos
      produces:
        - application/json
      responses:
        200:
          description: OK
  /pagamento:
    post:
      tags:
        - Pagamentos
      produces:
        - application/json
      responses:
        200:
          description: OK
    put:
      tags:
        - Pagamentos
      produces:
        - application/json
      responses:
        200:
          description: OK
  /pagamento/{id}:
    get:
      tags:
        - Pagamentos
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
          minimum: 1
          description: Parameter description in Markdown.
      responses:
        200:
          description: OK
    delete:
      tags:
        - Pagamentos
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
          minimum: 1
          description: Parameter description in Markdown.
      responses:
        200:
          description: OK

  /pedidos:
    get:
      tags:
        - Pedidos
      produces:
        - application/json
      responses:
        200:
          description: OK
  /pedido:
    post:
      tags:
        - Pedidos
      produces:
        - application/json
      responses:
        200:
          description: OK
    put:
      tags:
        - Pedidos
      produces:
        - application/json
      responses:
        200:
          description: OK
  /pedido/{id}:
    get:
      tags:
        - Pedidos
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
          minimum: 1
          description: Parameter description in Markdown.
      responses:
        200:
          description: OK
    delete:
      tags:
        - Pedidos
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
          minimum: 1
          description: Parameter description in Markdown.
      responses:
        200:
          description: OK

  /produtos:
    get:
      tags:
        - Produtos
      produces:
        - application/json
      responses:
        200:
          description: OK
  /produto:
    post:
      tags:
        - Produtos
      produces:
        - application/json
      responses:
        200:
          description: OK
    put:
      tags:
        - Produtos
      produces:
        - application/json
      responses:
        200:
          description: OK
  /produto/{id}:
    get:
      tags:
        - Produtos
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
          minimum: 1
          description: Parameter description in Markdown.
      responses:
        200:
          description: OK
    delete:
      tags:
        - Produtos
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
          minimum: 1
          description: Parameter description in Markdown.
      responses:
        200:
          description: OK